renamed type hints file
